move all resources into 1 dir
vs individual resource dirs
