+++ gunpla_api/resources/model.py
