+++ gunpla_api/resources/model_scale.py
