+++ gunpla_api/resources/series.py
